Cambios varios. Se agrega filtro a reportes.

+++ finalPwfe/src/app/productos/agregar-productos/agregar-productos.component.ts
@@ -11,6 +11,7 @@ export class AgregarProductosComponent implements OnInit {
 
   producto: Producto = new Producto();
   productoId: number = 0;
+  mensaje: String = "";
 
   constructor(private router: Router) { }
 
@@ -32,26 +33,45 @@ export class AgregarProductosComponent implements OnInit {
     }
   }
   saveProducto() {
-    debugger;
-    //Llega id actualizado para guardar
-    this.productoId =  this.getNewProductId();
-    this.producto.productoId = this.productoId;
+    if(this.producto.codigo == undefined){
+      this.mensaje = "--Introduzca un Codigo--";
+    }else{
+      if(this.producto.nombre == undefined){
+        this.mensaje = "--Introduzca un Nombre--";
+      }else{
+        if(this.producto.precioVenta == undefined){
+          this.mensaje = "--Introduzca un Precio--";
+        }else{
+          if(this.producto.existencia == undefined){
+            this.mensaje = "--Seleccione Existencia--";
+          }else{
+            //Llega id actualizado para guardar
+            this.productoId =  this.getNewProductId();
+            this.producto.productoId = this.productoId;
 
-    const datos =  localStorage.getItem('productos');
-    if (datos !== null) {
-      const productos = JSON.parse(datos);
-      productos.push(this.producto);
-      localStorage.setItem('productos', JSON.stringify(productos));
-      //Actualizar idProd
-      localStorage.setItem('productoId', JSON.stringify(this.productoId));
-    } else {
-      const productoArr = [];
-      productoArr.push(this.producto);
-      localStorage.setItem('productos', JSON.stringify(productoArr));
-      localStorage.setItem('productoId', JSON.stringify(1));
+            const datos =  localStorage.getItem('productos');
+            if (datos !== null) {
+              const productos = JSON.parse(datos);
+              productos.push(this.producto);
+              localStorage.setItem('productos', JSON.stringify(productos));
+              //Actualizar idProd
+              localStorage.setItem('productoId', JSON.stringify(this.productoId));
+            } else {
+              const productoArr = [];
+              productoArr.push(this.producto);
+              localStorage.setItem('productos', JSON.stringify(productoArr));
+              localStorage.setItem('productoId', JSON.stringify(1));
+            }
+            this.router.navigateByUrl('/productos');
+          }
+        }
+      }
     }
-    this.router.navigateByUrl('/productos');
+    debugger;
+  }
 
+  cancelar(){
+    this.router.navigateByUrl('/productos');
   }
 
 }
