+++ finalPwfe/src/app/productos/editar-productos/editar-productos.component.ts
@@ -11,6 +11,7 @@ export class EditarProductosComponent implements OnInit {
 
   producto: Producto = new Producto();
   productos: Producto[] = [];
+  mensaje: String = "";
   constructor(private route: ActivatedRoute, private router: Router) {
     this.route.params.subscribe((res)=> {
       this.producto.productoId = res['id']
@@ -33,19 +34,40 @@ export class EditarProductosComponent implements OnInit {
 
 
   updateProducto() {
-    debugger;
-    const datos =  localStorage.getItem('productos');
-    if (datos !== null) {
-      this.productos = JSON.parse(datos);
-      for(let i=0; i<this.productos.length; i++){
-        if( this.productos[i].productoId == this.producto.productoId){
-          this.productos[i] = this.producto;
+    if(this.producto.codigo == ""){
+      this.mensaje = "--Introduzca un Codigo--";
+    }else {
+      if (this.producto.nombre == "") {
+        this.mensaje = "--Introduzca un Nombre--";
+      } else {
+        if (this.producto.precioVenta == undefined) {
+          this.mensaje = "--Introduzca un Precio--";
+        } else {
+          if (this.producto.existencia == "") {
+            this.mensaje = "--Seleccione Existencia--";
+          } else {
+            debugger;
+            const datos = localStorage.getItem('productos');
+            if (datos !== null) {
+              this.productos = JSON.parse(datos);
+              for (let i = 0; i < this.productos.length; i++) {
+                if (this.productos[i].productoId == this.producto.productoId) {
+                  this.productos[i] = this.producto;
+                }
+              }
+              localStorage.setItem('productos', JSON.stringify(this.productos));
+            }
+            this.router.navigateByUrl('/productos');
+          }
         }
       }
-      localStorage.setItem('productos', JSON.stringify(this.productos));
     }
-    this.router.navigateByUrl('/productos');
 
+
+  }
+
+  cancelar(){
+    this.router.navigateByUrl('/productos');
   }
 
 }
